+++ aquadoggo/src/bus.rs
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use crate::service_manager::Sender;
+use p2panda_rs::entry::Entry;
+use p2panda_rs::operation::Operation;
+
+use crate::manager::Sender;
 
 /// Sender for cross-service communication bus.
 pub type ServiceSender = Sender<ServiceMessage>;
@@ -8,5 +11,6 @@ pub type ServiceSender = Sender<ServiceMessage>;
 /// Messages which can be sent on the communication bus.
 #[derive(Clone)]
 pub enum ServiceMessage {
-    // Nothing here to see yet ..!
+    /// New `Entry` with an `Operation` payload arrived at the node.
+    NewEntryAndOperation(Entry, Operation),
 }