+++ aquadoggo/src/materializer/tasks/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod dependency;
+mod reduce;
+mod schema;
+
+pub use dependency::dependency_task;
+pub use reduce::reduce_task;
+pub use schema::schema_task;