+++ aquadoggo/src/materializer/tasks/schema.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use crate::context::Context;
+use crate::materializer::worker::TaskResult;
+use crate::materializer::TaskInput;
+
+pub async fn schema_task(context: Context, input: TaskInput) -> TaskResult<TaskInput> {
+    Ok(None)
+}